+++ requirement_bazaar/src/main/java/de/rwth/dbis/acis/bazaar/service/BazaarService.java
@@ -114,7 +114,7 @@ public class BazaarService extends RESTService {
         getResourceConfig().register(CommentsResource.class);
         getResourceConfig().register(AttachmentsResource.class);
         getResourceConfig().register(UsersResource.class);
-        getResourceConfig().register(PersonalisationDataResource.class);
+        //getResourceConfig().register(PersonalisationDataResource.class);
         getResourceConfig().register(FeedbackResource.class);
     }