+++ requirement_bazaar/src/main/java/de/rwth/dbis/acis/bazaar/service/security/AuthorizationManager.java
@@ -81,6 +81,5 @@ public class AuthorizationManager {
         for (PrivilegeEnum privilege : privileges) {
             facade.createPrivilegeIfNotExists(privilege);
         }
-
     }
 }