+++ CHANGELOG.md
@@ -18,12 +18,15 @@ the [GitHub Release Page](https://github.com/rwth-acis/RequirementsBazaar/releas
 - Added endpoint to search for users by name or email [#90](https://github.com/rwth-acis/RequirementsBazaar/pull/90)
 - Categories, projects and requirements now have a `lastActivity`
   attribute [#91](https://github.com/rwth-acis/RequirementsBazaar/pull/91).
+- Categories, projects and requirements now have a `userContext` encapsuling the dynamic user related information (
+  permissions, votes, contribution) [#94](https://github.com/rwth-acis/RequirementsBazaar/pull/94).
+- If a user is a member of a project the respective role is now returned in the`projectRole` attribute of the
+  new `userContext` attribute [#94](https://github.com/rwth-acis/RequirementsBazaar/pull/94).
 
 ### Changed
 
 - Updated all dependencies, most notably las2peer 1.0.0 [#68](https://github.com/rwth-acis/RequirementsBazaar/pull/68)
-- Updated las2peer to 1.1.0 and thereby requiring Java
-  14) [#73](https://github.com/rwth-acis/RequirementsBazaar/pull/73)
+- Updated las2peer to 1.1.0 and thereby requiring Java 14 [#73](https://github.com/rwth-acis/RequirementsBazaar/pull/73)
 - Changed buildsystem to use gradle [#73](https://github.com/rwth-acis/RequirementsBazaar/pull/73)
 - Automatically generate jooq code from migration files at build
   time [#73](https://github.com/rwth-acis/RequirementsBazaar/pull/73)
@@ -48,6 +51,14 @@ the [GitHub Release Page](https://github.com/rwth-acis/RequirementsBazaar/releas
   las2peerid [#90](https://github.com/rwth-acis/RequirementsBazaar/pull/90)
 - Requirements no longer return the category objects in the `categories` attribute but a list of category
   ids [#91](https://github.com/rwth-acis/RequirementsBazaar/pull/91).
+- Vote direction can no longer be provided as a query
+  parameter [#94](https://github.com/rwth-acis/RequirementsBazaar/pull/94)
+- Moved user related information in categories, requirements and projects (isFollower/Developer/Contributor, userVoted)
+  into the new `userContext` [#94](https://github.com/rwth-acis/RequirementsBazaar/pull/94).
+
+### Removed
+
+- Personalisation endpoints [#94](https://github.com/rwth-acis/RequirementsBazaar/pull/94)
 
 ## [0.7.2] - 2017-10-25